analyzer/pyproject.toml
[tool.poetry]
name = "analyzer"
version = "0.1.0"
description = ""
authors = ["Константин <k1st.volkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
mysql-connector-python = "^8.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
